+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # chat21-ionic ver 3.0
 
+### 3.0.98-rc.1
+- changed: conversation.attributes.channel with conversation.attributes.request_channel to get the current convesation channel
+- bug-fixed: whatsapp template not enabled when agent is added to a whatsapp channel request
+
+### 3.0.97 in PROD
+
 ### 3.0.97-rc.5
 - added: whatsapp templates section to send an already configured whatsapp templete only in case of a whatsapp conversation channel type
 - added: channel input var in message-text-area component to manage the type of current conversation